+++ CHANGES.rst
@@ -5,6 +5,7 @@ OpenREM version history
 0.9.0b6??
 --------------------
 * `#714`_  Documentation: Add missing documentation about changing STATIC_URL if serving OpenREM in a virtual directory
+* `#688`_  Interface: Add possibility to apply known display name based on Device Observer UID (default: disabled)
 
 0.9.0b5 (2018-12-21)
 --------------------